Fix mask generator float32 bug and add local experiment script with results
The SAM2 automatic mask generator expects uint8 input but was receiving
float32, causing it to produce 0 candidate masks. Also adds a local
experiment script that benchmarks the few-shot pipeline against held-out
annotations, and documents results across model sizes and point densities.

+++ workers/annotations/sam2_fewshot_segmentation/entrypoint.py
@@ -407,7 +407,7 @@ def compute(datasetId, apiUrl, token, params):
         merged_image_rgb = ensure_rgb(merged_image)
 
         # Generate candidate masks with SAM2
-        candidate_masks = mask_generator.generate(merged_image_rgb.astype(np.float32))
+        candidate_masks = mask_generator.generate(merged_image_rgb)
         print(f"Frame {i + 1}: generated {len(candidate_masks)} candidate masks")
 
         # Filter candidates by similarity to training prototype
